Fix bug where server could crash after a restart due to prior connection
Restarting the server would remove all game instance objects. However,
if there were any clients still running from the previous instance of
the server, they would still have a reference to their socket. When the
client would attempt to perform an action, the server would try to
locate the game instance that the client socket was attached to.
There would not be an existing game instance and the server would crash
trying to carry out an operation on non-existing data.