Modify procedure for getting the list of games that can be joined
Previously, once a player clicked 'Join Game' the client would
send a request to the server every second to get the list of games
that could be joined. The server would then calculate this information
on the fly and send it back to the client. Repeat this for as many
clients as are interested in joining a game. It was not a very
efficient process.

Now a list of joinable games is tracked and modified as necessary
whenever a server action is triggered, such as creating a new game,
joining a game, or leaving a game. When a player clicks 'Join Game',
they get the current version of the list and are then subscribed to
any future changes to the list.