Add rotation animation to the player tokens
Previously, the game board was visually updated by deleting all of the
player tokens and redrawing them every time a change was made to the
board. This did not work with the proposed animation effects, because
redrawing a player token would abruptly reset the animation cycle.

Now the player tokens are retained until the board is specifically
cleared. Placing a new token draws only the new token in the specified
position without deleting all of the others as well.